+++ app/src/CheckList.tsx
@@ -12,20 +12,28 @@ export const CheckList = () => (
         <p>Only if connecting one or more biosynthetic gene clusters to molecules/molecular families:</p>
         <ul>
             <li>
-                Deposit the gene cluster in <a target="_blank" rel="noopener noreferrer" href="https://mibig.secondarymetabolites.org/">MIBiG</a> and have the accession number available</li>
+                Deposit the gene cluster in <a target="_blank" rel="noopener noreferrer" href="https://mibig.secondarymetabolites.org/">MIBiG</a> and have the accession number available
+            </li>
             <li>
-                Have the SMILES (textual representation of the chemical structure) of the metabolites available</li>
+                Have the SMILES (textual representation of the chemical structure) of the metabolites available
+            </li>
             <li>
-                Locate the MS2 scan or molecular family ID in <a target="_blank" rel="noopener noreferrer" href="https://gnps.ucsd.edu/">GNPS</a></li>
+                Locate the MS2 scan or molecular family ID in <a target="_blank" rel="noopener noreferrer" href="https://gnps.ucsd.edu/">GNPS</a>
+            </li>
         </ul>
         <p>Pro Tips:</p>
         <ul>
-            <li>Load the sample data to get an idea of how to fill out the form.</li>
+            <li>
+                Load the sample data to get an idea of how to fill out the form.
+            </li>
             <li>
                 Fill out the form in order from section 1 to 6.</li>
             <li>
                 To save your work, complete sections 1-4 for at least one sample and make at least one link in section 5, then press the Preview button, scroll down to the table created and press the Download button. You can later use the Upload button to upload this file to continue working. Validation errors may appear at the top of the page and must be resolved before you can download the file.
-        </li>
+            </li>
+            <li>
+                A file can also be uploaded at the bottom of the form. The file should adhere to the <a href="http://pairedomicsdata.bioinformatics.nl/schema.json">JSON schema</a>.
+            </li>
         </ul>
     </div>
 );